RPI/requirements.txt
adafruit-circuitpython-dht 
paho-mqtt 
smbus